fix: rename annotation -> adapter
