+++ src/main/kotlin/com/meetfootball/common/adapter/Persistence.kt
@@ -1,4 +1,4 @@
-package com.meetfootball.common.annotation
+package com.meetfootball.common.adapter
 
 import org.springframework.core.annotation.AliasFor
 import org.springframework.stereotype.Component
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component
 @Retention(AnnotationRetention.RUNTIME)
 @MustBeDocumented
 @Component
-annotation class PersistenceAdapter(
+annotation class Persistence(
     @get:AliasFor(annotation = Component::class)
     val value: String = "",
 )
